requirements.txt
lightning==2.5.0.post0
matplotlib==3.10.0
numpy==2.2.2
opencv_contrib_python==4.10.0.84
pycocotools==2.0.8
PyYAML==6.0.2
PyYAML==6.0.2
segmentation_models_pytorch==0.4.0
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
ultralytics==8.3.57
